refact: #55 move page to component
