+++ frontend/src/components/_mobile/mobile-search-page/MobileSearchPage.test.jsx
