+++ frontend/src/components/anonymous-feed/AnonymousFeed.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import PostList from '../components/posts/PostList';
-import { getPostsByType, appendPosts } from '../modules/post';
+import { getPostsByType, appendPosts } from '@modules/post';
+import PostList from '@components/posts/PostList';
 
 const AnonymousFeed = () => {
   const [target, setTarget] = useState(null);
