+++ frontend/src/components/friends-page/FriendsPage.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
-import { getFriendList } from '../modules/friend';
-import FriendItem from '../components/friends/FriendItem';
+import { getFriendList } from '@modules/friend';
+import FriendItem from '@components/friends/FriendItem';
 
 const FriendListWrapper = styled.div`
   padding: 16px;
