+++ frontend/src/components/friends-page/FriendsPage.test.jsx
