+++ frontend/src/components/post-detail/PostDetail.test.jsx
@@ -6,8 +6,8 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import PostDetail from './PostDetail';
-import PostItem from '../components/posts/PostItem';
-import QuestionItem from '../components/posts/QuestionItem';
+import PostItem from './posts/PostItem';
+import QuestionItem from './posts/QuestionItem';
 import rootReducer from '../modules';
 import { mockStore, mockStoreWithArticle } from '../mockStore';
 import 'jest-styled-components';
