+++ frontend/src/components/question-selection/QuestionSelection.jsx
@@ -2,12 +2,12 @@ import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import styled from 'styled-components';
-import { getSampleQuestions } from '../modules/question';
+import { getSampleQuestions } from '../../modules/question';
 import {
   postSelectedQuestions,
   skipOrCompleteSelectQuestions
-} from '../modules/user';
-import { CommonButton } from '../styles';
+} from '../../modules/user';
+import { CommonButton } from '../../styles';
 
 const QuestionsWrapper = styled.div`
   margin: 30px auto;
